save vrm instead of glb
